Added Reverse Processing to the Net
Added functions that let me get POSSIBLE inputs from a matrix of outputs. Updated the Sigmoid test and added an input/output test. Also updated the MNIST app to run with the optimal number of epochs and batch size and added a file that will eventually run an app to "imagine" by drawing inputs.